# pkg/api/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pkg/api/services.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Users"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/manager/api/v1/login": {
      "post": {
        "operationId": "Users_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiLoginResponse"
            }
          },
          "401": {
            "description": "You have provided no authorization token or the token provided is invalid.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiLoginRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manager/api/v1/me": {
      "get": {
        "operationId": "Users_GetMe",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetMeResponse"
            }
          },
          "401": {
            "description": "You have provided no authorization token or the token provided is invalid.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "operationId": "Users_UpdateMe",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUpdateMeResponse"
            }
          },
          "401": {
            "description": "You have provided no authorization token or the token provided is invalid.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUpdateMeRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manager/api/v1/me/device-token": {
      "post": {
        "operationId": "Users_UpdateDeviceToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUpdateDeviceTokenResponse"
            }
          },
          "401": {
            "description": "You have provided no authorization token or the token provided is invalid.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUpdateDeviceTokenRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manager/api/v1/search": {
      "post": {
        "operationId": "Users_SearchUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiSearchUserResponse"
            }
          },
          "401": {
            "description": "You have provided no authorization token or the token provided is invalid.",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiSearchUserRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "definitions": {
    "apiGetMeResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/apiUser"
        }
      }
    },
    "apiLoginRequest": {
      "type": "object"
    },
    "apiLoginResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/apiUser"
        }
      }
    },
    "apiPlayerColor": {
      "type": "string",
      "enum": [
        "BLUE",
        "RED",
        "YELLOW",
        "GREEN",
        "BLACK",
        "PURPLE",
        "ORANGE",
        "PINK",
        "NEUTRAL",
        "BRONZE"
      ],
      "default": "BLUE"
    },
    "apiSearchUserRequest": {
      "type": "object",
      "properties": {
        "search": {
          "type": "string"
        }
      }
    },
    "apiSearchUserResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiUserSettings"
          }
        }
      }
    },
    "apiUpdateDeviceTokenRequest": {
      "type": "object",
      "properties": {
        "deviceToken": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "apiUpdateDeviceTokenResponse": {
      "type": "object"
    },
    "apiUpdateMeRequest": {
      "type": "object",
      "properties": {
        "settings": {
          "$ref": "#/definitions/apiUserSettings"
        }
      }
    },
    "apiUpdateMeResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/apiUser"
        }
      }
    },
    "apiUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "settings": {
          "$ref": "#/definitions/apiUserSettings"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiUserSettings": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string"
        },
        "color": {
          "$ref": "#/definitions/apiPlayerColor"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "description": "Enter the token with the `Bearer: ` prefix, e.g. Bearer abcde12345.",
      "name": "Authorization",
      "in": "header"
    }
  }
}